Refactored `validatePostOperationResultsAndRecord()` from `ActiveActiveStoreIngestionTask` into `StorePartitionDataReceiver`. 🧹
